Add point cloud viewing script
This script reads ALL point clouds from the provided directory, and
given the integer list passed at the command line, will present those
point clouds for viewing.

This is not the nicest CLI calling convention to use, but it was the
easiest to whip together quickly.

To run this script:
```python3
python3 view-point-cloud.py <pcd_directory> [view-int-list] [voxel-scale-factor]
```